requirements.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
tensorflow
Dataset credits: https://www.kaggle.com/arjuntejaswi/plant-village